# pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/internal/loan-package-offers/expire": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Manual trigger expire loan offers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan package offer",
                    "admin",
                    "internal"
                ],
                "summary": "Manual trigger expire loan offers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.BaseResponse-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/loan-package-offers/{id}/assign-loan": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin assign loan id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan package offer",
                    "admin"
                ],
                "summary": "Admin assign loan id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_core_loanoffer_transport_http.AdminAssignLoanIdRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.BaseResponse-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/loan-package-offers/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin cancel loan package offer interest",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan package offer",
                    "admin"
                ],
                "summary": "Admin cancel loan package offer interest",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.BaseResponse-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/loan-package-offers/{id}/offline-updates": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get offline offer update history",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan package offer",
                    "admin"
                ],
                "summary": "Get offline offer update history",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.BaseResponse-array_financing-offer_internal_core_entity_OfflineOfferUpdate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create offline offer update",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan package offer",
                    "admin"
                ],
                "summary": "Create offline offer update",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "offerUpdate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_core_loanoffer_transport_http.CreateOfferUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.BaseResponse-financing-offer_internal_core_entity_OfflineOfferUpdate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/my-loan-package-offer": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Investor find loan offers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan package offer",
                    "investor"
                ],
                "summary": "Investor find loan offers",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "pageNumber",
                        "name": "page[number]",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "page[size]",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "symbol",
                        "name": "symbol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "assetType",
                        "name": "assetType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.BaseResponse-array_financing-offer_internal_core_entity_LoanPackageOffer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/my-loan-package-offer/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Investor get loan package offer by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan package offer",
                    "investor"
                ],
                "summary": "Investor get loan package offer by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.BaseResponse-financing-offer_internal_core_entity_LoanPackageOffer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/my-loan-package-offer/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Investor cancel loan package offer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan package offer",
                    "investor"
                ],
                "summary": "Investor cancel loan package offer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.BaseResponse-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "financing-offer_internal_core_entity.AssetType": {
            "type": "string",
            "enum": [
                "UNDERLYING",
                "DERIVATIVE"
            ],
            "x-enum-varnames": [
                "AssetTypeUnderlying",
                "AssetTypeDerivative"
            ]
        },
...
        "financing-offer_internal_core_entity.FlowType": {
            "type": "string",
            "enum": [
                "001",
                "002",
                "ONLINE",
                "OFFLINE"
            ],
            "x-enum-varnames": [
                "FLowTypeDnseOffline",
                "FlowTypeDnseOnline",
                "Online",
                "Offline"
            ]
        },
...
        "financing-offer_internal_core_entity.LoanPackageOffer": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "expiredAt": {
                    "type": "string"
                },
                "flowType": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.FlowType"
                },
                "id": {
                    "type": "integer"
                },
                "loanPackageOfferInterests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageOfferInterest"
                    }
                },
                "loanPackageRequest": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageRequest"
                },
                "loanPackageRequestId": {
                    "type": "integer"
                },
                "offeredBy": {
                    "type": "string"
                },
                "symbol": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.Symbol"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "financing-offer_internal_core_entity.LoanPackageOfferInterest": {
            "type": "object",
            "properties": {
                "assetType": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.AssetType"
                },
                "cancelledAt": {
                    "type": "string"
                },
                "cancelledBy": {
                    "type": "string"
                },
                "cancelledReason": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.CancelledReason"
                },
                "contractSize": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "feeRate": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "initialRate": {
                    "type": "number"
                },
                "interestRate": {
                    "type": "number"
                },
                "limitAmount": {
                    "type": "number"
                },
                "loanContract": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.LoanContract"
                },
                "loanId": {
                    "type": "integer"
                },
                "loanPackageOffer": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageOffer"
                },
                "loanPackageOfferId": {
                    "type": "integer"
                },
                "loanRate": {
                    "type": "number"
                },
                "scoreGroupInterestId": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageOfferInterestStatus"
                },
                "submissionSheetDetailId": {
                    "type": "integer"
                },
                "term": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "financing-offer_internal_core_entity.OfflineOfferUpdate": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "note": {
                    "type": "string"
                },
                "offerId": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.OfflineOfferUpdateStatus"
                }
            }
        },
        "financing-offer_internal_core_entity.OfflineOfferUpdateStatus": {
            "type": "string",
            "enum": [
                "PROCESSING",
                "REJECTED",
                "APPROVED"
            ],
            "x-enum-varnames": [
                "OfflineOfferUpdateStatusProcessing",
                "OfflineOfferUpdateStatusRejected",
                "OfflineOfferUpdateStatusApproved"
            ]
        },
...
        "financing-offer_internal_handler.BaseResponse-array_financing-offer_internal_core_entity_LoanPackageOffer": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageOffer"
                    }
                }
            }
        },
...
        "financing-offer_internal_handler.BaseResponse-array_financing-offer_internal_core_entity_OfflineOfferUpdate": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/financing-offer_internal_core_entity.OfflineOfferUpdate"
                    }
                }
            }
        },
...
        "financing-offer_internal_handler.BaseResponse-financing-offer_internal_core_entity_LoanPackageOffer": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageOffer"
                }
            }
        },
...
        "financing-offer_internal_handler.BaseResponse-financing-offer_internal_core_entity_OfflineOfferUpdate": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.OfflineOfferUpdate"
                }
            }
        },
...
        "financing-offer_internal_handler.BaseResponse-string": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                }
            }
        },
        "financing-offer_internal_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "internal_core_loanoffer_transport_http.AdminAssignLoanIdRequest": {
            "type": "object",
            "required": [
                "interestRate",
                "loanId"
            ],
            "properties": {
                "interestRate": {
                    "type": "number"
                },
                "loanId": {
                    "type": "integer"
                }
            }
        },
        "internal_core_loanoffer_transport_http.CreateOfferUpdateRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}